opendns/requirements.txt
appdirs==1.4.0
dnspython==1.15.0
packaging==16.8
pyparsing==2.1.10
pyperclip==1.5.27
six==1.10.0
